refactor: page controller 분리
