+++ src/main/java/roomescape/controller/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.controller.exception;
 
 public record ErrorResponse(String message) {
 }
