+++ src/main/java/roomescape/controller/web/AdminPageController.java
@@ -1,20 +1,10 @@
-package roomescape.controller;
+package roomescape.controller.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class PageController {
-
-    @GetMapping("/")
-    public String mainPage() {
-        return "index";
-    }
-
-    @GetMapping("/reservation")
-    public String reservationPage() {
-        return "reservation";
-    }
+public class AdminPageController {
 
     @GetMapping("/admin")
     public String adminPage() {
