+++ src/main/java/roomescape/controller/web/UserPageController.java
@@ -1,0 +1,18 @@
+package roomescape.controller.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UserPageController {
+
+    @GetMapping("/")
+    public String mainPage() {
+        return "index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservationPage() {
+        return "reservation";
+    }
+}
