+++ src/test/java/roomescape/controller/UserPageControllerTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-class PageControllerTest extends BaseControllerTest {
+class UserPageControllerTest extends BaseControllerTest {
 
     @Test
     @DisplayName("메인 페이지를 조회한다.")
